Refactor ts => js
