+++ README.md
@@ -1,11 +1,4 @@
 # Alviverdes.io
 
-- Executa o projeto no browser
-
-      npm run server
-
-- Aplica alterações dinâmicamente
-
-      npm run dev
 
 https://www.escoladnc.com.br/blog/como-subir-um-projeto-nodejs-para-producao-na-vercel/
