--- package.json
@@ -1,23 +1,0 @@
-{
-  "name": "alviverdes.io",
-  "version": "1.0.0",
-  "description": "Portal de notícias palmeirense",
-  "main": "webpack.config.js",
-  "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack --progress --watch",
-    "prod": "NODE_ENV=production webpack --progress --watch",
-    "server": "live-server --port=5000 --host=localhost src"
-  },
-  "author": "Leandro",
-  "license": "MIT",
-  "dependencies": {
-    "@supabase/supabase-js": "^2.44.4"
-  },
-  "devDependencies": {
-    "cross-env": "^7.0.3",
-    "live-server": "^1.2.2",
-    "ts-loader": "^9.5.1",
-    "webpack": "^5.93.0",
-    "webpack-cli": "^5.1.4"
-  }
-}
