+++ src/assets/styles/news.css
@@ -49,7 +49,7 @@ header .nav-link:hover {
 /* ===== MAIN ===== */
 main .card:hover {
   cursor: pointer;
-  border: solid 2px rgba(38, 97, 33, 1);
+  border: solid 1px rgba(38, 97, 33, 1);
 }
 
 .main-card-img {
