+++ src/index.html
@@ -285,7 +285,7 @@
     </footer>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script type="text/javascript" src="./dist/app/index.js"></script>
+    <script type="text/javascript" src="./js/index.js"></script>
 </body>
 
 </html>
