+++ src/pages/games.html
@@ -96,7 +96,7 @@
     </footer>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script type="text/javascript" src="../dist/app/news.js"></script>
+    <script type="text/javascript" src="../js/index.js"></script>
 </body>
 
 </html>
