+++ src/pages/news.html
@@ -315,7 +315,7 @@
     </footer>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script type="text/javascript" src="../dist/app/news.js"></script>
+    <script type="text/javascript" src="../js/news.js"></script>
 </body>
 
 </html>
