+++ src/pages/players.html
@@ -176,7 +176,7 @@
         </div>
     </footer>
 
-    <script src="../dist/app/players.js"></script>
+    <script src="../js/players.js"></script>
 </body>
 
 </html>
