--- tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es6",
-        "module": "commonjs",
-        "outDir": "src/dist",
-        "strict": true
-    },
-    "include": [
-        "src"
-    ]
-}
